Version number: 1.0.7.6; Optimize way to release memory

+++ Soul/src/scenarioG/FightFrame.java
@@ -14,22 +14,24 @@ public class FightFrame {
 		dialog.addKeyListener(new KeyAdapter() {
 			@Override
 			public void keyPressed(KeyEvent e) {
-				if(count == 0)
-					dialog.setTxt("Kill them!!");
-				else if(count == 1)
-					dialog.setAsideTxt("Try to kill demons!!");
-				else if(count == 2)
+				if(e.getKeyCode() == KeyEvent.VK_SPACE) {
+					if(count == 0)
+						dialog.setTxt("Kill them!!");
+					else if(count == 1)
+						dialog.setAsideTxt("Try to kill demons!!");
+					else if(count == 2)
 						dialog.setAsideTxt("(Click your sprite and select action you want)");
-				else if(count == 3)
-					dialog.setAsideTxt("(after action the sprite must wait)");
-				else if(count == 4)
-					dialog.setAsideTxt("(KILL ALL DEMONS)");
-				else if(count == 5)
-					dialog.setHintTxt("KILL!!!!");
-				else if(count == 6) {
-					dialog.dispose();
+					else if(count == 3)
+						dialog.setAsideTxt("(after action the sprite must wait)");
+					else if(count == 4)
+						dialog.setAsideTxt("(KILL ALL DEMONS)");
+					else if(count == 5)
+						dialog.setHintTxt("KILL!!!!");
+					else if(count == 6) {
+						dialog.dispose();
+					}
+					count++;
 				}
-				count++;
 			}
 		});
 	}
